% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jazzodata-package.R
\docType{data}
\name{esac_meta}
\alias{esac_meta}
\title{Essen Folk Song Collection (Jazzomat version) Metadata}
\format{
A tibble with 7289 rows and 14 variables:
\describe{
\item{id}{ESAC ID of the tune (not completely unique!)}
\item{collection}{Collection}
\item{esac_title}{Orignal EsAC title}
\item{region}{Region of origin}
\item{esac_key}{Annotated key (tonic)}
\item{unit}{Annotated minimal rhythmic unit}
\item{signature}{Annotated signature (possible more than one, space separate,  or FREE)}
\item{source}{Annotated key (tonic)}
\item{comment}{Comment section}
\item{cnr}{CNR section (with yet unknown meaning!)}
\item{tunefamily}{Tunefamily based on EsAC-Ids}
\item{text}{Annotated lyrics}
\item{melstring}{Original melody code}
\item{signature}{Signature(s) of solo}
}
}
\usage{
esac_meta
}
\description{
Metadata for the  Essen Folk Song Collection (EFSC or ESAC) with 7.825 tunes (as provided by the Jazzomat project).
}
\keyword{datasets}
