% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/jazzodata-package.R
\docType{data}
\name{esac_transforms}
\alias{esac_transforms}
\title{Essen Folk Song Collection (Jazzomat version) Transformations}
\format{
A tibble with 7289 rows and 14 variables:
\describe{
\item{id}{ESAC ID of the tune (not completely unique!)}
\item{durclass_abs_raw}{Raw output of duration classes (absolute mode)}
\item{durclass_rel_raw}{Raw output of duration classes (relative mode)}
\item{fuzzyint_raw}{Raw output (Fuzzy interval values over the full melody)}
\item{phrase_id_raw}{Sequence of event-wise phrase IDs}
\item{int_raw}{ist of semitone intervals}
\item{ioiclass_abs_raw}{Raw output of IOI classes of all melody notes (absolute mode)}
\item{ioiclass_rel_raw}{Raw output of IOI classes of all melody notes (relative mode)}
\item{mcm_48}{Position in metrical circle map with N=48 divisions}
\item{pc_raw}{Raw output (values for all melody notes)}
\item{pitch_raw}{Raw output (values of all melody notes)}
\item{tpc_raw}{Raw output of note-wise TPC values}
}

A tibble with 7289 rows and 14 variables:
\describe{
\item{id}{ESAC ID of the tune (not completely unique!)}
\item{durclass_abs_raw}{Raw output of duration classes (absolute mode)}
\item{durclass_rel_raw}{Raw output of duration classes (relative mode)}
\item{fuzzyint_raw}{Raw output (Fuzzy interval values over the full melody)}
\item{phrase_id_raw}{Sequence of event-wise phrase IDs}
\item{int_raw}{ist of semitone intervals}
\item{ioiclass_abs_raw}{Raw output of IOI classes of all melody notes (absolute mode)}
\item{ioiclass_rel_raw}{Raw output of IOI classes of all melody notes (relative mode)}
\item{mcm_48}{Position in metrical circle map with N=48 divisions}
\item{pc_raw}{Raw output (values for all melody notes)}
\item{pitch_raw}{Raw output (values of all melody notes)}
\item{tpc_raw}{Raw output of note-wise TPC values}
}
}
\usage{
esac_transforms

esac_transforms
}
\description{
Some basic transformations (also known as viewpoints) of the 7,825 tunes in the Essen Folk Song Database (EFSC or ESAC)

Some basic transformations (also known as viewpoints) of the 7,825 tunes in the Essen Folk Song Database (EFSC or ESAC)
}
\keyword{datasets}
