% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/jazzodata-package.R
\docType{package}
\name{jazzodata-package}
\alias{jazzodata}
\alias{jazzodata-package}
\title{jazzodata: Data sets from the Jazzomat project}
\description{
R data package containing some useful data sets derived from the Weimar Jazz Database, the Omnibook, and the Essen Folk Song Collection.

R data package containing some useful data sets derived from the Weimar Jazz Database, the Omnibook, and the Essen Folk Song Collection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/klausfrieler/jazzodata}
}


Useful links:
\itemize{
  \item \url{https://github.com/klausfrieler/jazzodata}
}

}
\author{
\strong{Maintainer}: Klaus Frieler \email{klaus.frieler@ae.mpg.de}

}
\keyword{internal}
