% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/jazzodata-package.R
\docType{data}
\name{omnibook_meta}
\alias{omnibook_meta}
\title{Omnibook Metadata}
\format{
A tibble with 56 rows and 33 variables:
\describe{
\item{signature}{Annotated signature (possible more than one, space separated, or FREE)}
\item{performer}{Performer (= Charlie Parker)}
\item{title}{Title}
\item{titleaddon}{Add-on for title}
\item{solopart}{Running number of a solo in a piece}
\item{full_title}{Full title (including solo part and title addon)}
\item{instrument}{Instrument of the soloist ( = as)}
\item{style}{Style of piece (= BEBOP)}
\item{avgtempo}{Avg. Tempo (bpm)}
\item{tempoclass}{Tempo class. One of SLOW, MEDIUM SLOW, MEDIUM, MEDIUM UP, UP or empty}
\item{rhythmfeel}{Basic rhythmic feel/groove style. One of TWOBEAT, SWING, BALLAD, LATIN, FUNK or empty}
\item{key}{Key or tonal center of the piece}
\item{filename_sv}{Filename of originating SV project file}
\item{status}{Status of transcription. Always FINAL}
\item{composer}{Composer of the piece}
\item{tonality_type}{Tonality type of the composition (= FUNCTIONAL)}
\item{genre}{Genre of the composition (= ORIGINAL)}
\item{form}{Form of the composition}
\item{signature}{Signature(s) of solo}
\item{mbzid}{Music Brainz ID (= NA)}
\item{recordingdate}{Recording date(s) of containing record}
\item{recordingyear}{Recording year of containing record}
\item{recordbib}{Bibliography of record}
\item{recordtitle}{Title of containing record}
\item{label}{Record label of containing record}
\item{solostart}{Always -1}
\item{soloend}{Always -1}
\item{lineup}{Line-up of track}
\item{chord_changes}{Chord changes as compact string}
\item{chorus_count}{Number of full choruses in the solo}
\item{harmony_template}{Underlying harmony template of composition}
\item{solo_time}{Always NA}
}

A tibble with 56 rows and 33 variables:
\describe{
\item{signature}{Annotated signature (possible more than one, space separated, or FREE)}
\item{performer}{Performer (= Charlie Parker)}
\item{title}{Title}
\item{titleaddon}{Add-on for title}
\item{solopart}{Running number of a solo in a piece}
\item{full_title}{Full title (including solo part and title addon)}
\item{instrument}{Instrument of the soloist ( = as)}
\item{style}{Style of piece (= BEBOP)}
\item{avgtempo}{Avg. Tempo (bpm)}
\item{tempoclass}{Tempo class. One of SLOW, MEDIUM SLOW, MEDIUM, MEDIUM UP, UP or empty}
\item{rhythmfeel}{Basic rhythmic feel/groove style. One of TWOBEAT, SWING, BALLAD, LATIN, FUNK or empty}
\item{key}{Key or tonal center of the piece}
\item{filename_sv}{Filename of originating SV project file}
\item{status}{Status of transcription. Always FINAL}
\item{composer}{Composer of the piece}
\item{tonality_type}{Tonality type of the composition (= FUNCTIONAL)}
\item{genre}{Genre of the composition (= ORIGINAL)}
\item{form}{Form of the composition}
\item{signature}{Signature(s) of solo}
\item{mbzid}{Music Brainz ID (= NA)}
\item{recordingdate}{Recording date(s) of containing record}
\item{recordingyear}{Recording year of containing record}
\item{recordbib}{Bibliography of record}
\item{recordtitle}{Title of containing record}
\item{label}{Record label of containing record}
\item{solostart}{Always -1}
\item{soloend}{Always -1}
\item{lineup}{Line-up of track}
\item{chord_changes}{Chord changes as compact string}
\item{chorus_count}{Number of full choruses in the solo}
\item{harmony_template}{Underlying harmony template of composition}
\item{solo_time}{Always NA}
}
}
\usage{
omnibook_meta

omnibook_meta
}
\description{
Metadata for the 56 solos in the  Charlie Parker Omnibook (CPO).
See \url{https://jazzomat.hfm-weimar.de/dbformat/glossary.html} for further information.

Metadata for the 56 solos in the  Charlie Parker Omnibook (CPO).
See \url{https://jazzomat.hfm-weimar.de/dbformat/glossary.html} for further information.
}
\keyword{datasets}
