% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/jazzodata-package.R
\docType{data}
\name{omnibook_transforms}
\alias{omnibook_transforms}
\title{Omnibook Transformations}
\format{
A tibble with 18,569 rows and 21 variables:
\describe{
\item{id}{Solo ID}
\item{beat1}{Raw vector of accents on the primary accent (first beat) of a bar}
\item{beat3}{Raw vector of accents on the secondary accent of a bar, if present, e.g. in on the 3rd beat in 4/4-measures (hence the name)}
\item{beatall}{Raw vector of accents on all beat positions in a bar}
\item{phrasbeg}{Raw vector of markers for phrase beginnings (if available)}
\item{phrasend}{Raw vector of markers for phrase endings (if available)}
\item{cdpcx_raw_all}{Raw output of note-wise CDPCX values including NC values}
\item{cpc_raw_all}{Raw output (CPC values of all melody notes, including undefined CPCs, set to -1)}
\item{durclass_abs_raw}{Raw output of duration classes (absolute mode)}
\item{durclass_rel_raw}{Raw output of duration classes (relative mode)}
\item{fuzzyint_raw}{Raw output (Fuzzy interval values over the full melody)}
\item{metrical_position}{Metrical positions of events (MP-dot-notation)}
\item{phrase_id_raw}{Sequence of event-wise phrase IDs}
\item{chords_raw}{Sequence of event-wise chord symbols (as annotated and normalized)}
\item{chord_types_raw}{Sequence of event-wise chord type labels (triad type + seventh)}
\item{int_raw}{ist of semitone intervals}
\item{ioiclass_abs_raw}{Raw output of IOI classes of all melody notes (absolute mode)}
\item{ioiclass_rel_raw}{Raw output of IOI classes of all melody notes (relative mode)}
\item{metrical_position}{Metrical positions (in  MPD-notation) of events}
\item{mcm_48}{Position in metrical circle map with N=48 divisions}
\item{pitch_raw}{Raw output (values of all melody notes)}
}

A tibble with 18,569 rows and 21 variables:
\describe{
\item{id}{Solo ID}
\item{beat1}{Raw vector of accents on the primary accent (first beat) of a bar}
\item{beat3}{Raw vector of accents on the secondary accent of a bar, if present, e.g. in on the 3rd beat in 4/4-measures (hence the name)}
\item{beatall}{Raw vector of accents on all beat positions in a bar}
\item{phrasbeg}{Raw vector of markers for phrase beginnings (if available)}
\item{phrasend}{Raw vector of markers for phrase endings (if available)}
\item{cdpcx_raw_all}{Raw output of note-wise CDPCX values including NC values}
\item{cpc_raw_all}{Raw output (CPC values of all melody notes, including undefined CPCs, set to -1)}
\item{durclass_abs_raw}{Raw output of duration classes (absolute mode)}
\item{durclass_rel_raw}{Raw output of duration classes (relative mode)}
\item{fuzzyint_raw}{Raw output (Fuzzy interval values over the full melody)}
\item{metrical_position}{Metrical positions of events (MP-dot-notation)}
\item{phrase_id_raw}{Sequence of event-wise phrase IDs}
\item{chords_raw}{Sequence of event-wise chord symbols (as annotated and normalized)}
\item{chord_types_raw}{Sequence of event-wise chord type labels (triad type + seventh)}
\item{int_raw}{ist of semitone intervals}
\item{ioiclass_abs_raw}{Raw output of IOI classes of all melody notes (absolute mode)}
\item{ioiclass_rel_raw}{Raw output of IOI classes of all melody notes (relative mode)}
\item{metrical_position}{Metrical positions (in  MPD-notation) of events}
\item{mcm_48}{Position in metrical circle map with N=48 divisions}
\item{pitch_raw}{Raw output (values of all melody notes)}
}
}
\usage{
omnibook_transforms

omnibook_transforms
}
\description{
Some basic transformation (also known as viewpoints) from the 56 solos in the Charlie Parker Omnibook.

Some basic transformation (also known as viewpoints) from the 56 solos in the Charlie Parker Omnibook.
}
\keyword{datasets}
