% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wjd_annotated_beat_tracks}
\alias{wjd_annotated_beat_tracks}
\title{Weimar Jazz Database annotated beat tracks}
\format{
A tibble with 132,325 rows and 17 variables:
#' \describe{
\item{onset}{Onset of beat}
\item{duration}{Beat duration}
\item{period}{Beats in a bar}
\item{division}{Division of beat}
\item{bar}{Bar number}
\item{beat}{beat in bar}
\item{signature}{Time signature}
\item{chord}{Chord annotations}
\item{form}{Forma label}
\item{chorus_id}{Id of chorus in solos}
\item{bass}{Extimated bass pitch at this beat}
\item{id}{Solo ID}
\item{chord_change}{Flag whether chord changes here.}
\item{melid}{Integer version of solo ID (1 --456)}
\item{beat_id}{Unique beat identfier (melid_bar_beat)}
\item{local_scale_degree}{Local scale degree of this chord in local key (as per algorithm)}
\item{local_key}{Local key context for this chord (as per algorithm)}
}
}
\usage{
wjd_annotated_beat_tracks
}
\description{
This data frame contains the beat tracks from the Weimar Jazz Database with extra local key annotations.
}
\keyword{datasets}
