% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wjd_features}
\alias{wjd_features}
\title{Weimar Jazz Database Scalar features}
\format{
A tibble with 45 rows and 169 variables:
\describe{
\item{id}{Solo ID}
\item{mean_length_arpeggio_ascending}{Mean length of ascending arpeggios}
\item{mean_length_arpeggio_descending}{Mean length of descending arpeggios}
\item{mean_length_arpeggio}{Mean length of arpeggios (mixed up/down thirds)}
\item{cdpcx_density_1}{Relative frequency CDPCX = '1'}
\item{cdpcx_density_2}{Relative frequency CDPCX = '2'}
\item{cdpcx_density_3}{Relative frequency CDPCX =  '3'}
\item{cdpcx_density_4}{Relative frequency CDPCX =  '4'}
\item{cdpcx_density_5}{Relative frequency CDPCX =  '5'}
\item{cdpcx_density_6}{Relative frequency CDPCX =  '6'}
\item{cdpcx_density_T}{Relative frequency CDPCX =  'T'}
\item{cdpcx_density_B}{Relative frequency CDPCX =  'B'}
\item{cdpcx_density_L}{Relative frequency CDPCX = 'L'}
\item{cdpcx_density_b2}{Relative frequency CDPCX =  '-' (b2)}
\item{cdpcx_density_b6}{Relative frequency CDPCX =  '\%' (b6)}
\item{recordingyear}{Recording year of containing record}
\item{mean_length_arpeggio_ascending}{Mean length of ascending arpeggios}
\item{mean_length_arpeggio_descending}{Mean length of descending arpeggios}
\item{mean_length_arpeggio}{Mean length of arpeggios (mixed up/down thirds)}
\item{cdpcx_density_1}{Relative frequency CDPCX = '1'}
\item{cdpcx_density_2}{Relative frequency CDPCX = '2'}
\item{cdpcx_density_3}{Relative frequency CDPCX =  '3'}
\item{cdpcx_density_4}{Relative frequency CDPCX =  '4'}
\item{cdpcx_density_5}{Relative frequency CDPCX =  '5'}
\item{cdpcx_density_6}{Relative frequency CDPCX =  '6'}
\item{cdpcx_density_T}{Relative frequency CDPCX =  'T'}
\item{cdpcx_density_B}{Relative frequency CDPCX =  'B'}
\item{cdpcx_density_L}{Relative frequency CDPCX = 'L'}
\item{cdpcx_density_b2}{Relative frequency CDPCX =  '-' (b2)}
\item{cdpcx_density_b6}{Relative frequency CDPCX =  '\%' (b13)}
\item{cdpcx_density_7}{Relative frequency CDPCX =  '7'}
\item{cdpcx_entropy}{Normalized entropy of CDPCX distribution}
\item{mean_length_chromatic_ascending}{Mean length of ascending chromatic passages}
\item{mean_length_chromatic_descending}{Mean length of descending chromatic passages}
\item{mean_length_chromatic_mixed}{Mean length of chromatic passages (up and down combined)}
\item{cpc_bigram_entropy}{Unnormalized Entropy of bigram distribution in bits}
\item{cpc_hist_density_00}{Relative frequency CPC = 0}
\item{cpc_hist_density_01}{Relative frequency CPC = 1}
\item{cpc_hist_density_02}{Relative frequency CPC = 2}
\item{cpc_hist_density_03}{Relative frequency CPC = 3}
\item{cpc_hist_density_04}{Relative frequency CPC = 4}
\item{cpc_hist_density_05}{Relative frequency CPC = 5}
\item{cpc_hist_density_06}{Relative frequency CPC = 6}
\item{cpc_hist_density_07}{Relative frequency CPC = 7}
\item{cpc_hist_density_08}{Relative frequency CPC = 8}
\item{cpc_hist_density_09}{Relative frequency CPC = 9}
\item{cpc_hist_density_10}{Relative frequency CPC = 10}
\item{cpc_hist_density_11}{Relative frequency CPC = 11}
\item{cpc_circ_mean_angle}{Angle of circular mean of CPC distribution}
\item{cpc_circ_var}{Circular variance of CPC distribution (=1-length of circular mean)}
\item{cpc_entropy}{Normalized entropy of CPC distribution}
\item{durclass_bigram_entropy}{Unnormalized Entropy of bigram distribution in bits}
\item{durclass_abs_hist_01_very_short}{Relative frequency very short}
\item{durclass_abs_hist_02_short}{Relative frequency short}
\item{durclass_abs_hist_03_medium}{Relative frequency medium}
\item{durclass_abs_hist_04_long}{Relative frequency long}
\item{durclass_abs_hist_05_very_long}{Relative frequency very long}
\item{durclass_rel_hist_01_very_short}{Relative frequency very short}
\item{durclass_rel_hist_02_short}{Relative frequency short}
\item{durclass_rel_hist_03_medium}{Relative frequency medium}
\item{durclass_rel_hist_04_long}{Relative frequency long}
\item{durclass_rel_hist_05_very_long}{Relative frequency very long}
\item{durclass_abs_entropy}{Entropy of absolute duration class distributions}
\item{durclass_rel_entropy}{Entropy of relative duration class distributions}
\item{durclass_mean_seg_len}{Mean run-length averaged over all classes}
\item{mean_seg_len_01_very_short}{Mean run-length for very short events}
\item{mean_seg_len_02_short}{Mean run-length for short events}
\item{mean_seg_len_03_medium}{Mean run-length for medium long events}
\item{mean_seg_len_04_long}{Mean run-length for long events}
\item{mean_seg_len_05_very_long}{Mean run-length for very long events}
\item{fuzzyint_bigram_entropy}{Unnormalized Entropy of bigram distribution in bits}
\item{fuzzyint_hist_09_big_jump_down}{Relative frequency big jump down (-4)}
\item{fuzzyint_hist_08_jump_down}{Relative frequency jump down (-3)}
\item{fuzzyint_hist_07_leap_down}{Relative frequency leap down ( -2)}
\item{fuzzyint_hist_06_step_down}{Relative frequency step down (-1)}
\item{fuzzyint_hist_05_repeat}{Relative frequency repetition (0)}
\item{fuzzyint_hist_04_step_up}{Relative frequency step up (+1)}
\item{fuzzyint_hist_03_leap_up}{Relative frequency leap up (+2)}
\item{fuzzyint_hist_02_jump_up}{Relative frequency jump up (+3)}
\item{fuzzyint_hist_01_big_jump_up}{Relative frequency big jump up (+4)}
\item{fuzzyint_entropy}{Normalized entropy of FUZZYINT distribution}
\item{art_median}{Median of articulation distribution}
\item{art_std}{Standard deviation of articulation distribution}
\item{art_range}{Range of of articulation distribution}
\item{event_density}{Event density (notes per seconds)}
\item{total_duration}{Total duration in seconds}
\item{metrical_event_density}{Event density (notes per bar). WARNING: Might not be working properly with meter changes}
\item{total_duration_bar}{Total duration in fractional bar units (e.g., 6 Quarter notes in 4/4 timing are 1.5 bars long)}
\item{number_notes}{Number of events}
\item{performer}{Performer}
\item{title}{Title}
\item{instrument}{Instrument of the soloist. One of ts, ts-c, as, bs, bsx, cbsx, ss, sss, cl, bcl, acl, tp, tpt, flgn, tb, fl, cor, ptp, frhn, ob, voc  or empty}
\item{style}{Style of piece. One of TRADITIONAL, SWING,  BEBOP, COOL, HARDBOP, POSTBOP, FREE, FUSION, OTHER, MIX  or empty}
\item{avgtempo}{Avg. Tempo (bpm)}
\item{rhythmfeel}{Basic rhythmic feel/groove style. One of TWOBEAT, SWING, BALLAD, LATIN, FUNK or empty}
\item{tonality_type}{Tonality type of the composition. One of FUNCTIONAL, MODAL, COLOR, FREE, BLUES, JAZZ-BLUES or empty}
\item{int_bigram_entropy}{Unnormalized Entropy of bigram distribution in bits}
\item{mean_length_chromatic_sequences}{Mean run-length of chromatic passages}
\item{ratio_chromatic_sequences}{Percentage of of chromatic passages}
\item{int_std}{Standard deviation of interval distribution}
\item{int_min}{Minimum of of interval distribution}
\item{int_range}{Range of of interval distribution}
\item{int_entropy}{Entropy of interval distribution (bits)}
\item{abs_int_mean}{Mean of absolute interval distribution}
\item{abs_int_range}{Range of absolute interval distribution}
\item{ioiclass_bigram_entropy}{Unnormalized Entropy of bigram distribution in bits}
\item{ioiclass_abs_hist_04_long}{Relative frequency long (absolute mode)}
\item{ioiclass_rel_hist_02_short}{Relative frequency short (relative  mode)}
\item{ioiclass_rel_hist_04_long}{Relative frequency long (relative mode)}
\item{ioiclass_rel_hist_05_very_long}{Relative frequency very long (relative mode)}
\item{ioiclass_abs_hist_01_very_short}{Relative frequency very short  (absolute mode)}
\item{ioiclass_abs_hist_02_short}{Relative frequency short  (absolute mode)}
\item{ioiclass_abs_hist_03_medium}{Relative frequency medium (absolute mode)}
\item{ioiclass_abs_hist_05_very_long}{Relative frequency very long (absolute mode)}
\item{ioiclass_rel_hist_01_very_short}{Relative frequency very short (relative  mode)}
\item{ioiclass_rel_hist_03_medium}{Relative frequency medium (relative mode)}
\item{ioiclass_abs_entropy}{Entropy of absolute inter-onset interval class distributions}
\item{ioiclass_rel_entropy}{Entropy of relative inter-onset interval class distributions}
\item{aic_mean_seg_len}{Mean run-length averaged over all classes (absolute mode)}
\item{aic_mean_seg_len_01_very_short}{Mean run-length for very short events (absolute mode)}
\item{aic_mean_seg_len_02_short}{Mean run-length for short events (absolute mode)}
\item{aic_mean_seg_len_03_medium}{Mean run-length for medium long events (absolute mode)}
\item{aic_mean_seg_len_04_long}{Mean run-length for long events (absolute mode)}
\item{aic_mean_seg_len_05_very_long}{Mean run-length for very long events (absolute mode)}
\item{ric_mean_seg_len}{Mean run-length averaged over all classes (relative mode)}
\item{ric_mean_seg_len_01_very_short}{Mean run-length for very short events (relative mode)}
\item{ric_mean_seg_len_02_short}{Mean run-length for short events (relative mode)}
\item{ric_mean_seg_len_03_medium}{Mean run-length for medium long events (relative mode)}
\item{ric_mean_seg_len_04_long}{Mean run-length for long events  (relative mode)}
\item{ric_mean_seg_len_05_very_long}{Mean run-length for very long events (relative mode)}
\item{mcm_mean_angle}{Angle of circular mean of MCM distribution}
\item{mcm_var}{Circular variance of MCM distribution (=1-length of circular mean)}
\item{mcm_entropy}{Normalized entropy of MCM distribution}
\item{syncopicity}{Ratio of syncopated events to non-syncopated events}
\item{metric_complexity}{Combined metric complexity after Frieler, in preparation}
\item{metric_complexity_division}{Division-based metric complexity after Frieler, in preparation}
\item{metric_complexity_compression}{Compression-based metric complexity after Frieler, in preparation}
\item{parsons_bigram_entropy}{Unnormalized Entropy of bigram distribution in bits}
\item{parsons_hist_descending}{Relative frequency for down (-1)}
\item{parsons_hist_constant}{Relative frequency for repeat (0)}
\item{parsons_hist_ascending}{Relative frequency for up (+1)}
\item{parsons_entropy}{Normalized entropy of Parson's Code distribution}
\item{mean_length_seq_ascending}{Mean length of ascending passages}
\item{mean_length_seq_descending}{Mean length of descending passages}
\item{mean_length_seq_repitition}{Mean length of repetitions}
\item{recordingyear}{Recording year of containing record}
\item{ratio_ascending_descending}{Ratio of ascending to descending intervals}
\item{mean_number_of_unique_pc}{Average number of different pitch classes per phrase}
\item{pc_circ_var}{Circular variance of PC distribution (=1-length of circular mean)}
\item{pc_entropy}{Normalized entropy of PC distribution}
\item{pitch_entropy}{Entropy of PITCH distribution (bits)}
\item{f0_mod_range_cents}{Extracted modulation range (cents)}
\item{f0_mod_freq}{Extracted modulation range (Hz)}
\item{f0_median_dev}{Extracted median deviation from nominal 12TET pitch (cents)}
\item{loudness_max}{Maximum loudness values per tone}
\item{loudness_med}{Median loudness values per tone}
\item{loudness_sd}{Standard deviation of loudness per tone}
\item{loudness_rel_peak_pos}{Relative position of maximum loudness value, duration of tone equals 1}
\item{loudness_temp_cent}{Normed temporal centroid of loudness curve per tone, duration of tone equals 1}
\item{loudness_s2b}{Ratio of solo loudness (median) to backing track loudness (median)}
\item{pitch_median}{Median of PITCH distribution}
\item{pitch_std}{Standard deviation of PITCH distribution}
\item{pitch_range}{Range of of PITCH distribution}
\item{extrema_ratio}{Percentage of extremal points}
\item{mean_length_step_ascending}{Mean length of ascending step passages}
\item{mean_length_step_descending}{Mean length of descending step passages}
\item{mean_length_step}{Mean length of step passages (up and down steps)}
\item{nPVI_dur}{Normalized pairwise variability index}
\item{nPVI_ioi}{Normalized pairwise variability index}
}
}
\usage{
wjd_features
}
\description{
Most of available scalar features from the Weimar Jazz Database (WJD), including the most relevant metadata.
 See \url{https://jazzomat.hfm-weimar.de/melospy/transformations.html} for further information.
}
\keyword{datasets}
