% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jazzodata-package.R
\docType{data}
\name{wjd_tone}
\alias{wjd_tone}
\title{Weimar Jazz Database tone-based sound features}
\format{
A tibble with 200,809 rows and 13 variables:
\describe{
\item{id}{Solo ID}
\item{modulation}{Annonated modulation per tone}
\item{modulation_short}{Annonated modulation per tone (short label)}
\item{f0_mod_range_cents}{Extracted modulation range (cents)}
\item{f0_mod_freq}{Extracted modulation range (Hz)}
\item{f0_median_dev}{Extracted median deviation from nominal 12TET pitch (cents)}
\item{loudness_max}{Maximum loudness values per tone}
\item{loudness_med}{Median loudness values per tone}
\item{loudness_sd}{Standard deviation of loudness per tone}
\item{loudness_rel_peak_pos}{Relative position of maximum loudness value, duration of tone equals 1}
\item{loudness_temp_cent}{Normed temporal centroid of loudness curve per tone, duration of tone equals 1}
\item{loudness_s2b}{Ratio of solo loudness (median) to backing track loudness (median)}
\item{recordingyear}{Recording year of containing record}
}

A tibble with 184,015 rows and 30 variables:
\describe{
\item{id}{Solo ID}
\item{cdpcx}{Extebded chordal diationic pitch classes}
\item{cpc}{Chordal diationic pitch classes}
\item{bar}{Measure number}
\item{beat}{Beat number in metrical annotations}
\item{division}{Beat division in tatums }
\item{mcm_48}{Metrical circle map position (0-47)}
\item{period}{Number of beats in current bar}
\item{tatum}{Tatum position under beat of current event}
\item{pc}{Pitch class (C4 = 0)}
\item{pitch}{MIDI-Pitch  (C4 = 60)}
\item{melid}{Melid ( = numerical form od Solo ID), based on originalö 456 solos}
\item{beat_id}{Beat ID, unique enumeration of all WJD beats, format <melid>_<bar_number>_<beatnumber>}
\item{form_name}{Form name of current form part}
\item{chord}{Current chord}
\item{local_key}{Local key according to harmonic analysis}
\item{local_scale_degree}{Roman step annotation of current chord according to harmonic analysis}
\item{tonic}{Tonic of current chord (note name or NA)}
\item{key_pc}{Pitch class of current local key (0 to 11 and NA)}
\item{key}{Overall annotated key of the composition}
\item{genre}{Genre of the composition. One of TRADITIONAL, BLUES, GREAT AMERICAN SONGBOOK, WORMS, ORIGINAL, RIFF or empty}
\item{performer}{Performer of the solo)}
\item{title}{Title}
\item{recordingyear}{Recording year of containing record}
\item{rhythmfeel}{Basic rhythmic feel/groove style. One of TWOBEAT, SWING, BALLAD, LATIN, FUNK or empty}
\item{style}{Style of piece. One of TRADITIONAL, SWING,  BEBOP, COOL, HARDBOP, POSTBOP, FREE, FUSION, OTHER, MIX  or empty}
\item{tempoclass}{Tempo class. One of SLOW, MEDIUM SLOW, MEDIUM, MEDIUM UP, UP or empty}
\item{tonality_type}{Tonality type of the composition. One of FUNCTIONAL, MODAL, COLOR, FREE, BLUES, JAZZ-BLUES or empty}
}
}
\description{
Loudness (resp. intensity) related values for single tone events have been extracted from jazz solos by means of
sophisticated MIR techniques and are stored in the WJD
database, where five different values are available: Maximum, median and standard deviation of loudness,
relative position of loudness peak and temporal centroid. Here, only a meaningful subset is included.
See \url{https://jazzomat.hfm-weimar.de/melospy/transformations.html} for further information.

This data frame contains the most important features and meta-data for tonal analysis. Specifically, it contains a local
scale degree annotation created with the corresponding functions from \code{parkR}-package.
This also means, it contain fewer solos (429) than the WJD (456 solos), as not all solo have annotated chords that allow such an harmonic
analysis
}
