% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/jazzodata-package.R
\docType{data}
\name{wjd_tpc}
\alias{wjd_tpc}
\title{Weimar Jazz Database tonality based features}
\format{
A tibble with 184,015 rows and 30 variables:
\describe{
\item{id}{Solo ID}
\item{cdpcx}{Extebded chordal diationic pitch classes}
\item{cpc}{Chordal diationic pitch classes}
\item{bar}{Measure number}
\item{beat}{Beat number in metrical annotations}
\item{division}{Beat division in tatums }
\item{mcm_48}{Metrical circle map position (0-47)}
\item{period}{Number of beats in current bar}
\item{tatum}{Tatum position under beat of current event}
\item{pc}{Pitch class (C4 = 0)}
\item{pitch}{MIDI-Pitch  (C4 = 60)}
\item{melid}{Melid ( = numerical form od Solo ID), based on originalö 456 solos}
\item{beat_id}{Beat ID, unique enumeration of all WJD beats, format <melid>_<bar_number>_<beatnumber>}
\item{form_name}{Form name of current form part}
\item{chord}{Current chord}
\item{local_key}{Local key according to harmonic analysis}
\item{local_scale_degree}{Roman step annotation of current chord according to harmonic analysis}
\item{tonic}{Tonic of current chord (note name or NA)}
\item{key_pc}{Pitch class of current local key (0 to 11 and NA)}
\item{key}{Overall annotated key of the composition}
\item{genre}{Genre of the composition. One of TRADITIONAL, BLUES, GREAT AMERICAN SONGBOOK, WORMS, ORIGINAL, RIFF or empty}
\item{performer}{Performer of the solo)}
\item{title}{Title}
\item{recordingyear}{Recording year of containing record}
\item{rhythmfeel}{Basic rhythmic feel/groove style. One of TWOBEAT, SWING, BALLAD, LATIN, FUNK or empty}
\item{style}{Style of piece. One of TRADITIONAL, SWING,  BEBOP, COOL, HARDBOP, POSTBOP, FREE, FUSION, OTHER, MIX  or empty}
\item{tempoclass}{Tempo class. One of SLOW, MEDIUM SLOW, MEDIUM, MEDIUM UP, UP or empty}
\item{tonality_type}{Tonality type of the composition. One of FUNCTIONAL, MODAL, COLOR, FREE, BLUES, JAZZ-BLUES or empty}
}

A tibble with 184,015 rows and 30 variables:
\describe{
\item{id}{Solo ID}
\item{cdpcx}{Extebded chordal diationic pitch classes}
\item{cpc}{Chordal diationic pitch classes}
\item{bar}{Measure number}
\item{beat}{Beat number in metrical annotations}
\item{division}{Beat division in tatums }
\item{mcm_48}{Metrical circle map position (0-47)}
\item{period}{Number of beats in current bar}
\item{tatum}{Tatum position under beat of current event}
\item{pc}{Pitch class (C4 = 0)}
\item{pitch}{MIDI-Pitch  (C4 = 60)}
\item{melid}{Melid ( = numerical form od Solo ID), based on originalö 456 solos}
\item{beat_id}{Beat ID, unique enumeration of all WJD beats, format <melid>_<bar_number>_<beatnumber>}
\item{form_name}{Form name of current form part}
\item{chord}{Current chord}
\item{local_key}{Local key according to harmonic analysis}
\item{local_scale_degree}{Roman step annotation of current chord according to harmonic analysis}
\item{tonic}{Tonic of current chord (note name or NA)}
\item{key_pc}{Pitch class of current local key (0 to 11 and NA)}
\item{key}{Overall annotated key of the composition}
\item{genre}{Genre of the composition. One of TRADITIONAL, BLUES, GREAT AMERICAN SONGBOOK, WORMS, ORIGINAL, RIFF or empty}
\item{performer}{Performer of the solo)}
\item{title}{Title}
\item{recordingyear}{Recording year of containing record}
\item{rhythmfeel}{Basic rhythmic feel/groove style. One of TWOBEAT, SWING, BALLAD, LATIN, FUNK or empty}
\item{style}{Style of piece. One of TRADITIONAL, SWING,  BEBOP, COOL, HARDBOP, POSTBOP, FREE, FUSION, OTHER, MIX  or empty}
\item{tempoclass}{Tempo class. One of SLOW, MEDIUM SLOW, MEDIUM, MEDIUM UP, UP or empty}
\item{tonality_type}{Tonality type of the composition. One of FUNCTIONAL, MODAL, COLOR, FREE, BLUES, JAZZ-BLUES or empty}
}
}
\usage{
wjd_tpc

wjd_tpc
}
\description{
This data frame contains the most important features and meta-data for tonal analysis. Specifically, it contains a local
scale degree annotation created with the corresponding functions from \code{parkR}-package.
This also means, it contain fewer solos (444) than the WJD (456 solos), as not all solo have annotated chords that allow such an harmonic
analysis

This data frame contains the most important features and meta-data for tonal analysis. Specifically, it contains a local
scale degree annotation created with the corresponding functions from \code{parkR}-package.
This also means, it contain fewer solos (444) than the WJD (456 solos), as not all solo have annotated chords that allow such an harmonic
analysis
}
\keyword{datasets}
