% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jazzodata-package.R
\docType{data}
\name{wjd_transforms}
\alias{wjd_transforms}
\title{Weimar Jazz Database Transformations}
\format{
A tibble with 200,809 rows and 17 variables:
\describe{
\item{id}{Solo ID}
\item{cdpcx_raw}{Raw output of note-wise CDPCX values}
\item{cpc_raw}{Raw output (CPC values of all melody notes)}
\item{durclass_abs_raw}{Raw output of duration classes (absolute mode)}
\item{durclass_rel_raw}{Raw output of duration classes (relative mode)}
\item{fuzzyint_raw}{Raw output (Fuzzy interval values over the full melody)}
\item{articulation_raw}{Ratios of durations to interonset-intervals}
\item{phrase_id_raw}{Sequence of event-wise phrase IDs}
\item{chords_raw}{Sequence of event-wise chord symbols (as annotated and normalized)}
\item{int_raw}{ist of semitone intervals}
\item{ioiclass_abs_raw}{Raw output of IOI classes of all melody notes (absolute mode)}
\item{ioiclass_rel_raw}{Raw output of IOI classes of all melody notes (relative mode)}
\item{mcm_48}{Position in metrical circle map with N=48 divisions}
\item{pc_raw}{Raw output (values for all melody notes)}
\item{pitch_raw}{Raw output (values of all melody notes)}
}
}
\usage{
wjd_transforms
}
\description{
Some basic transformations (also known as viewpoints) from the Weimar Jazz Database (WJD). See \url{https://jazzomat.hfm-weimar.de/melospy/melospy.html} for further information.
}
