% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jazzodata-package.R
\docType{data}
\name{wjd_tone}
\alias{wjd_tone}
\title{Weimar Jazz Database tone-based sound features}
\format{
A tibble with 200,809 rows and 13 variables:
\describe{
\item{id}{Solo ID}
\item{modulation}{Annonated modulation per tone}
\item{modulation_short}{Annonated modulation per tone (short label)}
\item{f0_mod_range_cents}{Extracted modulation range (cents)}
\item{f0_mod_freq}{Extracted modulation range (Hz)}
\item{f0_median_dev}{Extracted median deviation from nominal 12TET pitch (cents)}
\item{loudness_max}{Maximum loudness values per tone}
\item{loudness_med}{Median loudness values per tone}
\item{loudness_sd}{Standard deviation of loudness per tone}
\item{loudness_rel_peak_pos}{Relative position of maximum loudness value, duration of tone equals 1}
\item{loudness_temp_cent}{Normed temporal centroid of loudness curve per tone, duration of tone equals 1}
\item{loudness_s2b}{Ratio of solo loudness (median) to backing track loudness (median)}
\item{recordingyear}{Recording year of containing record}
}
}
\description{
Loudness (resp. intensity) related values for single tone events have been extracted from jazz solos by means of
sophisticated MIR techniques and are stored in the WJD
database, where five different values are available: Maximum, median and standard deviation of loudness,
relative position of loudness peak and temporal centroid. Here, only a meaningful subset is included.
See \url{https://jazzomat.hfm-weimar.de/melospy/transformations.html} for further information.
}
